requirements.txt
asgiref==3.7.2
boto3==1.34.46
botocore==1.34.46
Django==4.2.10
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environ==1.0
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2024.1
s3transfer==0.10.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==1.26.18
